// generated-client/api.ts
export interface CategoryRequest {
    'name': string;
    'description'?: string;
    'difficulty': CategoryRequestDifficultyEnum;
    'orderIndex'?: number;
    'isActive'?: boolean;
}

export const CategoryRequestDifficultyEnum = {
    Beginner: 'BEGINNER',
    Intermediate: 'INTERMEDIATE',
    Advanced: 'ADVANCED'
} as const;

export type CategoryRequestDifficultyEnum = typeof CategoryRequestDifficultyEnum[keyof typeof CategoryRequestDifficultyEnum];
// ...
export class CategoriesApi extends BaseAPI {
    /**
     * 
     * @summary Create a new category
     * @param {CategoryRequest} categoryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public createCategory(categoryRequest: CategoryRequest, options?: RawAxiosRequestConfig) {
        return CategoriesApiFp(this.configuration).createCategory(categoryRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Delete category
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public deleteCategory(id: number, options?: RawAxiosRequestConfig) {
        return CategoriesApiFp(this.configuration).deleteCategory(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get all categories
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getAllCategories(options?: RawAxiosRequestConfig) {
        return CategoriesApiFp(this.configuration).getAllCategories(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get category by ID
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getCategoryById(id: number, options?: RawAxiosRequestConfig) {
        return CategoriesApiFp(this.configuration).getCategoryById(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Update category
     * @param {number} id 
     * @param {CategoryRequest} categoryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateCategory(id: number, categoryRequest: CategoryRequest, options?: RawAxiosRequestConfig) {
        return CategoriesApiFp(this.configuration).updateCategory(id, categoryRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class HealthApi extends BaseAPI {
    /**
     * 
     * @summary Health check endpoint
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public healthCheck(options?: RawAxiosRequestConfig) {
        return HealthApiFp(this.configuration).healthCheck(options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class SystemApi extends BaseAPI {
    /**
     * 
     * @summary API root endpoint
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getApiRoot(options?: RawAxiosRequestConfig) {
        return SystemApiFp(this.configuration).getApiRoot(options).then((request) => request(this.axios, this.basePath));
    }
}